LPD-15281 Transform AnonymousUserConfigurationRetriever into a util class, rename it to AnonymousUserConfigurationRetrieverUtil and apply to usages
