+++ modules/apps/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/model/listener/UserModelListener.java
@@ -9,7 +9,7 @@ import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.user.associated.data.web.internal.configuration.AnonymousUserConfigurationRetriever;
+import com.liferay.user.associated.data.web.internal.util.AnonymousUserConfigurationRetrieverUtil;
 
 import java.io.IOException;
 
@@ -39,8 +39,9 @@ public class UserModelListener extends BaseModelListener<User> {
 	private void _deleteAnonymousUserConfiguration(long companyId, long userId)
 		throws InvalidSyntaxException, IOException {
 
-		Configuration configuration = _anonymousUserConfigurationRetriever.get(
-			_configurationAdmin, companyId, userId);
+		Configuration configuration =
+			AnonymousUserConfigurationRetrieverUtil.get(
+				_configurationAdmin, companyId, userId);
 
 		if (configuration == null) {
 			return;
@@ -48,10 +49,6 @@ public class UserModelListener extends BaseModelListener<User> {
 
 		configuration.delete();
 	}
-
-	@Reference
-	private AnonymousUserConfigurationRetriever
-		_anonymousUserConfigurationRetriever;
 
 	@Reference
 	private ConfigurationAdmin _configurationAdmin;
