+++ modules/apps/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/util/AnonymousUserConfigurationRetrieverUtil.java
@@ -1,24 +1,24 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.user.associated.data.web.internal.configuration;
+package com.liferay.user.associated.data.web.internal.util;
+
+import com.liferay.user.associated.data.web.internal.configuration.AnonymousUserConfiguration;
 
 import java.io.IOException;
 
 import org.osgi.framework.InvalidSyntaxException;
 import org.osgi.service.cm.Configuration;
 import org.osgi.service.cm.ConfigurationAdmin;
-import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Drew Brokke
  */
-@Component(service = AnonymousUserConfigurationRetriever.class)
-public class AnonymousUserConfigurationRetriever {
+public class AnonymousUserConfigurationRetrieverUtil {
 
-	public Configuration get(
+	public static Configuration get(
 			ConfigurationAdmin configurationAdmin, long companyId)
 		throws InvalidSyntaxException, IOException {
 
@@ -29,7 +29,7 @@ public class AnonymousUserConfigurationRetriever {
 					_getFactoryPid(), companyId)));
 	}
 
-	public Configuration get(
+	public static Configuration get(
 			ConfigurationAdmin configurationAdmin, long companyId, long userId)
 		throws InvalidSyntaxException, IOException {
 
@@ -40,7 +40,7 @@ public class AnonymousUserConfigurationRetriever {
 					_getFactoryPid(), companyId, userId)));
 	}
 
-	private Configuration _get(Configuration[] configurations) {
+	private static Configuration _get(Configuration[] configurations) {
 		if (configurations == null) {
 			return null;
 		}
@@ -48,7 +48,7 @@ public class AnonymousUserConfigurationRetriever {
 		return configurations[0];
 	}
 
-	private String _getFactoryPid() {
+	private static String _getFactoryPid() {
 		return AnonymousUserConfiguration.class.getName() + ".scoped";
 	}
